Hardens sidebar action telemetry against review findings
- Guards the context-menu telemetry emit in the command wrapper so a
  telemetry failure can never prevent the command handler from running
  (right-clicking Delete Branch must work even if the emit path throws).
- Excludes remote branches (`gitlens:branch+remote`, the leaves nested
  under the Remotes panel) from action resolution: the Branches
  panel/metric is local-only and the inline path never emitted for
  them, so both surfaces now symmetrically leave remote-branch actions
  to `graph/command` instead of mis-attributing 24 context-menu
  commands to `branchAction`.
- Restores action×location coherence for the two inline-tracked branch
  compares (`compareWithHead`, `compareWithWorking`) by resolving them
  on the context-menu surface too.
- Merges the duplicated inline and context-menu command→action maps
  into one shared per-item-type table (`graphSidebarActionTelemetry.ts`,
  renamed from graphSidebarContextMenuTelemetry.ts) consumed by both
  surfaces, so a rename on one side cannot silently fragment the metric.
- Extracts the inline re-stamp into `markSidebarInlineInvocation` and
  tests the mark→gate coupling; documents the executeCommand routing
  invariant at the dispatch site.

(#5477)
